feat: update to latest better auth version

+++ src/routes/api/devices/delete_device/+server.ts
@@ -1,5 +1,4 @@
 import { json } from '@sveltejs/kit';
-import { auth } from '$lib/server/auth';
 import { db } from '$lib/server/db';
 import { eq, and } from 'drizzle-orm';
 import {
@@ -21,13 +20,9 @@ import { join } from 'path';
 
 import { env } from '$env/dynamic/private';
 
-export async function DELETE(event) {
+export async function DELETE({ locals, url }) {
   // Check if the user is authenticated
-  const session = await auth.api.getSession({
-    headers: event.request.headers
-  });
-
-  if (!session) {
+  if (!locals.user) {
     return json({ error: 'Unauthorized' }, { status: 401 });
   }
 
@@ -35,7 +30,7 @@ export async function DELETE(event) {
   const lastDeletedTime = await db
     .select({ lastDeletedTime: lastActionTimes.lastDeletedTime })
     .from(lastActionTimes)
-    .where(eq(lastActionTimes.userId, session.user.id))
+    .where(eq(lastActionTimes.userId, locals.user.id))
     .get();
 
   if (lastDeletedTime && lastDeletedTime.lastDeletedTime) {
@@ -54,12 +49,12 @@ export async function DELETE(event) {
     await db
       .update(lastActionTimes)
       .set({ lastDeletedTime: currentTime })
-      .where(eq(lastActionTimes.userId, session.user.id));
+      .where(eq(lastActionTimes.userId, locals.user.id));
   } else {
-    await db.insert(lastActionTimes).values({ userId: session.user.id }).onConflictDoNothing();
+    await db.insert(lastActionTimes).values({ userId: locals.user.id }).onConflictDoNothing();
   }
 
-  const deviceId = event.url.searchParams.get('id');
+  const deviceId = url.searchParams.get('id');
   if (!deviceId) {
     return json({ message: 'Device ID is required' }, { status: 400 });
   }
@@ -68,7 +63,7 @@ export async function DELETE(event) {
     const device = await db
       .select()
       .from(userDevices)
-      .where(and(eq(userDevices.id, parseInt(deviceId)), eq(userDevices.userId, session.user.id)))
+      .where(and(eq(userDevices.id, parseInt(deviceId)), eq(userDevices.userId, locals.user.id)))
       .get();
 
     if (!device) {
@@ -76,31 +71,35 @@ export async function DELETE(event) {
     }
 
     await db.transaction(async (tx) => {
+      if (!locals.user) {
+        return json({ error: 'Unauthorized' }, { status: 401 });
+      }
+
       await tx.delete(userDevices).where(eq(userDevices.id, parseInt(deviceId)));
 
       // Check if any other devices exist for each parameter
       if (device.cpu !== null && device.cpu !== undefined) {
-        await deleteOrphans(tx, cpus, device.cpu, session.user.id, 'cpu');
+        await deleteOrphans(tx, cpus, device.cpu, locals.user.id, 'cpu');
       }
 
       if (device.gpu !== null && device.gpu !== undefined) {
-        await deleteOrphans(tx, gpus, device.gpu, session.user.id, 'gpu');
+        await deleteOrphans(tx, gpus, device.gpu, locals.user.id, 'gpu');
       }
 
       if (device.memory !== null && device.memory !== undefined) {
-        await deleteOrphans(tx, memory, device.memory, session.user.id, 'memory');
+        await deleteOrphans(tx, memory, device.memory, locals.user.id, 'memory');
       }
 
       if (device.storage !== null && device.storage !== undefined) {
-        await deleteOrphans(tx, storage, device.storage, session.user.id, 'storage');
+        await deleteOrphans(tx, storage, device.storage, locals.user.id, 'storage');
       }
 
       if (device.os !== null && device.os !== undefined) {
-        await deleteOrphans(tx, os, device.os, session.user.id, 'os');
+        await deleteOrphans(tx, os, device.os, locals.user.id, 'os');
       }
 
       if (device.brand !== null && device.brand !== undefined) {
-        await deleteOrphans(tx, brands, device.brand, session.user.id, 'brand');
+        await deleteOrphans(tx, brands, device.brand, locals.user.id, 'brand');
       }
 
       // Delete device's image folder
